educa/requirements_version1.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.0.9
django-braces==1.15.0
django-embed-video==1.4.9
idna==3.10
pillow==10.3.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
